Cleared Visual Studio x64 warnings

+++ src/crypto/MessageDigest.cpp
@@ -259,7 +259,7 @@ namespace esapi
         throw EncryptionException(std::string("Internal error: ") + ex.what());
       }
 
-    return req;
+    return (unsigned int)req;
   }
 
   /**
